Description string parsing issue fixed I discovered that if the description section of the html had any time of special formatting, even just a bold character, BeautifulSoup would break that into multiple sections, causing the descriptions for many commands to get improperly truncated.
The solution currently is to capture the text immediately after "Command is undoable, queryable, and not editable", by the string "Return value".  I am not a fan of using such simple logic as this assumes that "Return value" never occurs within the description until the actual Return value header section.  I attempted to find Return value section using more accurate methods such as scanning for BeautifulSoup4.Tag header sections and breaking at that point.  To simply searching for "\nReturn value\n" as a more specific occurrence of this text.  But while that worked with some HTML docs, there would consistently be others that fell through.  So in lieu of a perfect solution, this seems to work well, and certainly better than the current iteration.